Fix calendar acl not showing stored category, selectAccount_ro widget expects value should be array or type of string

+++ calendar/inc/class.calendar_uiforms.inc.php
@@ -2335,7 +2335,7 @@ class calendar_uiforms extends calendar_ui
 		$n = 1;
 		foreach($this->bo->get_cat_rights() as $Lcat_id => $data)
 		{
-			$cat_id = (int)substr($Lcat_id,1);
+			$cat_id = substr($Lcat_id,1);
 			$row = array(
 				'cat_id' => $cat_id,
 				'add' => array(),
